+++ lib/scala/cli/src/main/scala/org/enso/cli/arguments/Application.scala
@@ -1,4 +1,4 @@
-package org.enso.cli
+package org.enso.cli.arguments
 
 import cats.data.NonEmptyList
 import org.enso.cli.internal.Parser
@@ -69,6 +69,9 @@ class Application[Config](
     * Runs the application logic. Parses the top level options and depending on
     * its result, possibly runs a command or a plugin.
     *
+    * If a plugin is run, this function does not return (the application exits
+    * with the exit code returned by the plugin).
+    *
     * @return either a list of errors encountered when parsing the options or
     *         [[Unit]] if it succeeded
     */
@@ -113,67 +116,23 @@ class Application[Config](
   def renderHelp(): String = combinedOpts.topLevelHelp(Seq(commandName))
 }
 
-/**
-  * A plugin manager that handles finding and running plugins.
-  */
-trait PluginManager {
-
-  /**
-    * Tries to run a given plugin with provided arguments.
-    *
-    * It never returns - either it exits with the plugin's exit code or throws
-    * an exception if it was not possible to run the plugin.
-    *
-    * @param name name of the plugin
-    * @param args arguments that should be passed to it
-    */
-  def runPlugin(name: String, args: Seq[String]): Nothing
-
-  /**
-    * Returns whether the plugin of the given `name` is available in the system.
-    */
-  def hasPlugin(name: String): Boolean
-
-  /**
-    * Lists names of plugins found in the system.
-    */
-  def pluginsNames(): Seq[String]
-
-  /**
-    * Lists names and short descriptions of plugins found on the system.
-    */
-  def pluginsHelp(): Seq[CommandHelp]
-}
-
-/**
-  * Defines the behaviour of parsing top-level application options.
-  *
-  * @tparam Config type of configuration that is passed to commands
-  */
-sealed trait TopLevelBehavior[+Config]
-object TopLevelBehavior {
-
-  /**
-    * If top-level options return a value of this class, the application should
-    * continue execution. The provided value of `Config` should be passed to the
-    * executed commands.
-    *
-    * @param withConfig the configuration that is passed to commands
-    * @tparam Config type of configuration that is passed to commands
-    */
-  case class Continue[Config](withConfig: Config)
-      extends TopLevelBehavior[Config]
-
-  /**
-    * If top-level options return a value of this class, it means that the
-    * top-level options have handled the execution and commands should not be
-    * parsed further. This can be useful to implement top-level options like
-    * `--version`.
-    */
-  case object Halt extends TopLevelBehavior[Nothing]
-}
-
 object Application {
+
+  /**
+    * Helper constructor for [[Application]].
+    *
+    * @param commandName default name of the application for use in commands
+    * @param prettyName pretty name of the application for use in text
+    * @param helpHeader short description of the application included at the
+    *                   top of the help text
+    * @param topLevelOpts the top-level options that are used to parse a global
+    *                     config which is passed to every command; can also be
+    *                     used to execute different bahavior than commands; see
+    *                     [[TopLevelBehavior]] for more information
+    * @param commands a sequence of commands supported by the application
+    * @param pluginManager a plugin manager for resolving non-native command
+    *                      extensions
+    */
   def apply[Config](
     commandName: String,
     prettyName: String,
@@ -191,6 +150,21 @@ object Application {
       Some(pluginManager)
     )
 
+  /**
+    * Helper constructor for [[Application]].
+    *
+    * Creates an application without plugin support.
+    *
+    * @param commandName default name of the application for use in commands
+    * @param prettyName pretty name of the application for use in text
+    * @param helpHeader short description of the application included at the
+    *                   top of the help text
+    * @param topLevelOpts the top-level options that are used to parse a global
+    *                     config which is passed to every command; can also be
+    *                     used to execute different bahavior than commands; see
+    *                     [[TopLevelBehavior]] for more information
+    * @param commands a sequence of commands supported by the application
+    */
   def apply[Config](
     commandName: String,
     prettyName: String,
@@ -207,6 +181,18 @@ object Application {
       None
     )
 
+  /**
+    * Helper constructor for [[Application]].
+    *
+    * Creates an application without any top-level options and without plugin
+    * support.
+    *
+    * @param commandName default name of the application for use in commands
+    * @param prettyName pretty name of the application for use in text
+    * @param helpHeader short description of the application included at the
+    *                   top of the help text
+    * @param commands a sequence of commands supported by the application
+    */
   def apply(
     commandName: String,
     prettyName: String,
