+++ lib/scala/cli/src/main/scala/org/enso/cli/arguments/IllegalOptsStructure.scala
@@ -1,0 +1,7 @@
+package org.enso.cli.arguments
+
+/**
+  * Exception that is reported when Opts are combined in an illegal way.
+  */
+case class IllegalOptsStructure(message: String, cause: Throwable = null)
+    extends RuntimeException(message, cause)
