+++ lib/scala/cli/src/main/scala/org/enso/cli/arguments/TopLevelBehavior.scala
@@ -1,0 +1,30 @@
+package org.enso.cli.arguments
+
+object TopLevelBehavior {
+
+  /**
+    * If top-level options return a value of this class, the application should
+    * continue execution. The provided value of `Config` should be passed to the
+    * executed commands.
+    *
+    * @param withConfig the configuration that is passed to commands
+    * @tparam Config type of configuration that is passed to commands
+    */
+  case class Continue[Config](withConfig: Config)
+      extends TopLevelBehavior[Config]
+
+  /**
+    * If top-level options return a value of this class, it means that the
+    * top-level options have handled the execution and commands should not be
+    * parsed further. This can be useful to implement top-level options like
+    * `--version`.
+    */
+  case object Halt extends TopLevelBehavior[Nothing]
+}
+
+/**
+  * Defines the behaviour of parsing top-level application options.
+  *
+  * @tparam Config type of configuration that is passed to commands
+  */
+sealed trait TopLevelBehavior[+Config]
