+++ lib/scala/cli/src/main/scala/org/enso/cli/internal/opts/OptionalParameter.scala
@@ -1,6 +1,6 @@
 package org.enso.cli.internal.opts
 
-import org.enso.cli.{Argument, OptsParseError}
+import org.enso.cli.arguments.{Argument, OptsParseError}
 
 class OptionalParameter[A: Argument](
   name: String,
@@ -33,7 +33,7 @@ class OptionalParameter[A: Argument](
     value = OptsParseError.combineWithoutDuplicates(
       value,
       Argument[A].read(newValue),
-      s"Multiple values for parameter $name."
+      s"Multiple values for parameter `$name`."
     )
   }
