+++ lib/scala/cli/src/main/scala/org/enso/cli/internal/opts/OptsMap.scala
@@ -1,7 +1,7 @@
 package org.enso.cli.internal.opts
 
+import org.enso.cli.arguments.{Opts, OptsParseError}
 import org.enso.cli.internal.ParserContinuation
-import org.enso.cli.{Opts, OptsParseError}
 
 class OptsMap[A, B](a: Opts[A], f: A => B) extends Opts[B] {
   override private[cli] def flags              = a.flags
