+++ lib/scala/cli/src/main/scala/org/enso/cli/internal/opts/OptsProduct.scala
@@ -1,7 +1,7 @@
 package org.enso.cli.internal.opts
 
+import org.enso.cli.arguments.{IllegalOptsStructure, Opts, OptsParseError}
 import org.enso.cli.internal.ParserContinuation
-import org.enso.cli.{IllegalOptsStructure, Opts, OptsParseError}
 
 class OptsProduct[A, B](lhs: Opts[A], rhs: Opts[B]) extends Opts[(A, B)] {
   override private[cli] def flags      = lhs.flags ++ rhs.flags
